Excel Update
Bot can now read in products from an excel spreadsheet. Much faster system as the user needs to only put in the desired website and the index of the product they're looking to buy. The spreadsheet contains the information the bot needs: the exact name of the product along with the most profitable shoe size.